stat
statistica grossa

+++ src/main/java/com/example/skiplegday/model/Data.java
@@ -1,6 +1,6 @@
 package com.example.skiplegday.model;
 
-public record Data(String date, int punteggio) implements Comparable<Data>{
+public record Data(String date, double punteggio) implements Comparable<Data>{
 	@Override
 	public int compareTo(Data o) {
 		return date.compareTo(o.date);
